Extract tt-kmd-lib to its own target
