+++ device/api/umd/device/pcie/silicon_tlb_handle.hpp
@@ -8,8 +8,9 @@
 #include <cstdint>
 
 #include "umd/device/pcie/tlb_handle.hpp"
-#include "umd/device/tt_kmd_lib/tt_kmd_lib.h"
 #include "umd/device/types/tlb.hpp"
+
+struct tt_tlb_t;
 
 namespace tt::umd {
