+++ device/pcie/pci_device.cpp
@@ -29,10 +29,10 @@
 #include <vector>
 
 #include "tracy.hpp"
+#include "tt-kmd-lib/pci_ids.h"
+#include "tt-kmd-lib/tt_kmd_lib.h"
 #include "umd/device/arch/architecture_implementation.hpp"
-#include "umd/device/pcie/pci_ids.h"
 #include "umd/device/pcie/silicon_tlb_handle.hpp"
-#include "umd/device/tt_kmd_lib/tt_kmd_lib.h"
 #include "umd/device/types/arch.hpp"
 #include "umd/device/utils/error.hpp"
 #include "umd/device/utils/kmd_versions.hpp"
